refactor: definitions moved ConfigData.h -> ConfigData.cpp
